test: migrate SerialAPICommandMachine tests to AVA
